requirements.txt
# Django requirements
Django==5.0.2
Pillow==10.2.0
djangorestframework==3.14.0



# Additional dependencies
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5 